Added Stars

+++ src/nl/joey/universe/entity/Coordinates.scala
@@ -1,6 +1,6 @@
 package nl.joey.universe.entity
 
-case class Coordinates(var x: Float, var y: Float, var z: Float){
+case class Coordinates(var x: Double, var y: Double, var z: Double){
   override def toString: String = {
     s"[X: $x, Y: $y, Z: $z]"
   }
@@ -10,6 +10,6 @@ case class Coordinates(var x: Float, var y: Float, var z: Float){
 
 object Coordinates{
   def empty(): Coordinates = {
-    Coordinates(0f,0f,0f)
+    Coordinates(0,0,0)
   }
 }
