+++ src/nl/joey/universe/entity/FormulaVariables.scala
@@ -1,15 +1,15 @@
 package nl.joey.universe.entity
 
-case class FormulaVariables(var a: Float,
-                            var e: Float,
-                            var I: Float,
-                            var L: Float,
-                            var w: Float,
-                            var o: Float,
-                            var W: Float,
-                            var M: Float,
-                            var EccAnomaly: Float,
-                            var n: Float){
+case class FormulaVariables(var a: Double,
+                            var e: Double,
+                            var I: Double,
+                            var L: Double,
+                            var w: Double,
+                            var o: Double,
+                            var W: Double,
+                            var M: Double,
+                            var EccAnomaly: Double,
+                            var n: Double){
   override def toString: String = {
     s"a:$a, e:$e, I:$I, L:$L, w:$w, o:$o, W:$W, M:$M, E:$EccAnomaly, n:$n"
   }
